requirements.txt
requests 
mwxml 
mwparserfromhell
tqdm
goldenverba
python-dotenv